+++ src/lightkurve/time.py
@@ -1,51 +1,46 @@
-"""Adds the BKJD and BTJD time format for use by Astropy's `Time` object."""
-from astropy.time.formats import TimeNumeric, day_frac
+"""Adds the BKJD and BTJD time format for use by Astropy's `Time` object.
+
+Caution: AstroPy time objects make a distinction between a time's format
+(e.g. ISO, JD, MJD) and its scale (e.g. UTC, TDB).  This can be confusing
+because the acronym "BTJD" refers both to a format (TJD) and to a scale (TDB).
+
+Note: the classes below derive from an AstroPy meta class which will automatically
+register the formats for use in AstroPy Time objects.
+"""
+from astropy.time.formats import TimeFromEpoch
 
 
-class TimeBKJD(TimeNumeric):
+class TimeBKJD(TimeFromEpoch):
     """
-    Barycentric Kepler Julian Date time format.
-    This represents the number of days since January 1, 2009 12:00:00 UTC.
-    BKJD is the format in which times are recorded in Kepler data products.
-    See Section 2.3.2 in the Kepler Archive Manual for details.
+    Barycentric Kepler Julian Date (BKJD): days since JD 2454833.0.
+
+    For example, 0 in BTJD is noon on January 1, 2009.
+
+    BKJD is the format in which times are recorded in data products from
+    NASA's Kepler Space Telescope, where it is always given in the
+    Barycentric Dynamical Time (TDB) scale by convention.
     """
-
-    name = "bkjd"
-    BKJDREF = 2454833  # Barycentric Kepler Julian Date offset
-
-    def set_jds(self, val1, val2):
-        self._check_scale(self._scale)  # Validate scale.
-        jd1, jd2 = day_frac(val1, val2)
-        jd1 += self.BKJDREF
-        self.jd1, self.jd2 = day_frac(jd1, jd2)
-
-    def to_value(self, **kwargs):
-        jd1 = self.jd1 - self.BKJDREF
-        jd2 = self.jd2
-        return super().to_value(jd1=jd1, jd2=jd2, **kwargs)
-
-    value = property(to_value)
+    name = 'bkjd'
+    unit = 1.0
+    epoch_val = 2454833
+    epoch_val2 = None
+    epoch_scale = 'tdb'
+    epoch_format = 'jd'
 
 
-class TimeBTJD(TimeNumeric):
+class TimeBTJD(TimeFromEpoch):
     """
-    Barycentric TESS Julian Date time format.
-    This represents the number of days since JD 2457000.0.
-    BTJD is the format in which times are recorded in TESS data products.
+    Barycentric TESS Julian Date (BTJD): days since JD 2457000.0.
+
+    For example, 0 in BTJD is noon on December 8, 2014.
+
+    BTJD is the format in which times are recorded in data products from
+    NASA's Transiting Exoplanet Survey Satellite (TESS), where it is
+    always given in the Barycentric Dynamical Time (TDB) scale by convention.
     """
-
-    name = "btjd"
-    BTJDREF = 2457000  # Barycentric TESS Julian Date offset
-
-    def set_jds(self, val1, val2):
-        self._check_scale(self._scale)  # Validate scale.
-        jd1, jd2 = day_frac(val1, val2)
-        jd1 += self.BTJDREF
-        self.jd1, self.jd2 = day_frac(jd1, jd2)
-
-    def to_value(self, **kwargs):
-        jd1 = self.jd1 - self.BTJDREF
-        jd2 = self.jd2
-        return super().to_value(jd1=jd1, jd2=jd2, **kwargs)
-
-    value = property(to_value)
+    name = 'btjd'
+    unit = 1.0
+    epoch_val = 2457000
+    epoch_val2 = None
+    epoch_scale = 'tdb'
+    epoch_format = 'jd'